cloud/pyproject.toml
[project]
name = "otel-health-cloud"
version = "0.1.0"
description = "OpenTelemetry community health metrics — push to Grafana Cloud via OTLP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[tool.uv]
dev-dependencies = []
